contributors/rodneyrehm/incoming/css3-transitions/pseudo-elements-001.html

<!DOCTYPE html>
<html>
    <head>
        <meta charset="utf-8">
    	<meta http-equiv="content-type" content="text/html; charset=utf-8">
        <title>CSS Test: Transitioning Pseudo Elemeents</title>
        <meta name="assert" content="Test checks that transitions are run on pseudo elements">
        <link rel="help" href="http://www.w3.org/TR/css3-transitions/#transition-property-property">
        <link rel="help" href="http://www.w3.org/TR/CSS21/generate.html#before-after-content">
        <link rel="help" href="http://www.w3.org/TR/css3-content/">
        <link rel="author" title="Rodney Rehm" href="http://rodneyrehm.de/en/">
        <meta name="flags" content="dom">
        
        <script src="/resources/testharness.js" type="text/javascript"></script>
        <script src="/resources/testharnessreport.js" type="text/javascript"></script>
        
        <script src="./support/helper.js" type="text/javascript"></script>
        <script src="./support/runAsyncHarness.js" type="text/javascript"></script>
        <script src="./support/generalTest.js" type="text/javascript"></script>
        <script src="./support/properties.js" type="text/javascript"></script>
    </head>
    <body>
        <!-- required by testharnessreport.js -->
        <div id="log"></div>
        <!-- elements used for testing -->
        <div id="fixture">
            <div id="container">
                <div id="transition">Hello World</div>
            </div>
        </div>

        <script>
            // see README.md for an explanation of how this test suite works
            
            // this test takes its time, give it 5 minutes to run
            setup({timeout: 300000});
            // init dom
            domFixture('#fixture');
            
            var tests = [
                {
                    name: "transition padding-left on :before",
                    pseudo: 'before',
                    property: 'padding-left',
                    flags: {},
                    from: {'padding-left': '1px', 'content': '""'},
                    to: {'padding-left': '10px'}
                }, {
                    name: "transition padding-left on :after",
                    pseudo: 'after',
                    property: 'padding-left',
                    flags: {},
                    from: {'padding-left': '1px', 'content': '""'},
                    to: {'padding-left': '10px'}
                }, {
                    name: "transition padding-left on :before, changing content",
                    pseudo: 'before',
                    property: 'padding-left',
                    flags: {},
                    from: {'padding-left': '1px', 'content': '"1"'},
                    to: {'padding-left': '10px', 'content': '"2"'}
                }, {
                    name: "transition padding-left on :after, changing content",
                    pseudo: 'after',
                    property: 'padding-left',
                    flags: {},
                    from: {'padding-left': '1px', 'content': '"1"'},
                    to: {'padding-left': '10px', 'content': '"2"'}
                }
            ];
            
            // missing support for requestAnimationFrame forces us to
            // run animations at a slower pace, so setTimeout has a
            // chance of being invoked before TransitionEnd
            var duration = getRequestAnimationFrame() ? '0.1s' : '0.5s';

            runAsyncHarness({
                // array of test data
                tests: tests,
                // milliseconds to wait before calling teardown and ending test
                duration: parseFloat(duration) * 1000,
                // prepare individual test
                setup: generalTest.setup,
                // cleanup after individual test
                teardown: generalTest.teardown,
                // run actual test, assertions can be used here!
                test: function(data, options) {
                    /*
                        // generated CSS:
                        
                        #fixture {}
                        #container {}
                        #container.to {}
                        #container.transition {}
                        #transition {}
                        #transition:after
                        {
                            padding-left:1px;
                            content:"";
                        }
                        #transition.transition:after
                        {
                            transition:all 0.1s linear 0s;
                        }
                        #transition.to:after
                        {
                            padding-left:10px;
                        }
                    */
                    var styles = {};
                    styles['#fixture'] = {};
                    styles['#container'] = data.parentStyle;
                    styles['#container.to'] = {};
                    styles['#container.transition'] = {};
                    styles['#transition'] = {};
                    styles['#transition:' + data.pseudo.name] = data.from;
                    styles['#transition.transition:' + data.pseudo.name] = {transition: 'all ' + duration + ' linear 0s'};
                    styles['#transition.to:' + data.pseudo.name] = data.to;

                    // prepare styles
                    generalTest.setStyle(data, styles);
                    // make sure values differ, if they don't, the property could most likely not be parsed
                    assert_not_equals(data.transition.from, data.transition.to, "initial and target values may not match");
                    // collect computed values during transition
                    generalTest.startValueCollection(data);
                    // kick off the transition
                    generalTest.startTransition(data);
                    
                    // make sure we didn't get the target value immediately.
                    // If we did, there wouldn't be a transition!
                    var current = data.pseudo.computedStyle(data.property);
                    assert_not_equals(current, data.transition.to, "must not be target value after start");
                },
                // run tests immediately before teardown, assertions can be used here!
                concluding: function(data) {
                    // Not testing events, as they aren't specified yet
                    // some time in the future this may be specified as follows:
                    // TransitionEnd event is to be triggered on the pseudo's host element,
                    // adding the Event.pseudoElement = "before"; attribute for differentiation
                    data.test.step(generalTest.assertIntermediateValuesFunc(data, 'pseudo'));
                },
                
                // called once all tests are done
                done: generalTest.done
            });
        </script>
    </body>
</html>
